Fix: allow IPv6 addresses in WARC-IP-Address
After fixing my IPv6 on my main server, I noticed AT was defaulting to IPv6 for Telegram. Taking a look at the WARC, they were adding "WARC-IP-Address: 2001:67c:4e8:1033:4:100:0:a".

Per the WARC 1.1 specification: "<per section 2.2 of RFC4291>". Reading section 2.2.1 of RFC4291, we appear to be perfectly in spec with the IP addresses returned.

Double checking with the same URL with this change on Zeno, we get "WARC-IP-Address: 2001:67c:4e8:1033:4:100:0:a", the exact same, RFC4291 compliant IPv6 address.

In conclusion: we have the same behavior as wget-at and as far as I can tell, compliant with RFC4291 as requested by the WARC 1.1 spec.