DRAFT: write temporary files to disk to hopefully avoid OOM
Add WARCTempDir to our custom HTTP client, currently automatically configured, but we could add some configuration options in the future.

Change GetSHA1FromFile to support the headers being added separately from the actual file.

dialer.go: copy a limited number of bytes, right now 10K, if we get exactly 10K, we probably hit a limit, from there we want to split out the header, create a temporary file, add the content we previously got and the rest of the pipe.

Change up how we are reading the HTTP response depending on how we are writing it to the WARC file (either file, or .Content).

Deduplication works in the exact same way, but we have to remove files when we no longer need them.

write.go:
Remove the temporary files after they're written with a defer. Check if we sent the headers through r.Content, if we did, read it all, then use it for various calculations like Content-Length and Block-Digest.

After that, we use those headers, if they exist, to write to the actual WARC file.

After testing and coding this for the last week, this _appears_ to work. The code is a bit rough, but I've played back sites with Transfer-Encoding, video files, images, etc. Everything I've seen so far has the correct Payload-Digest (block-digest is a bit harder to validate..., but seems correct), and has perfectly valid warcs (as far as I can tell!), while being written to temporary files.