Some small proxy fixes (#5)
* Fix: WARC-Target-URI when used with HTTP/S proxies.

* fix: Panic at the end of Zeno run.

If a warc file doesn't have any records inside of it, it will panic, because there's no file on the filesystem. I currently don't believe this to be panic worthy, but it is obviously important that a file we expect to be there is missing.

* Fix: the old solution wasn't really a "fix" and the problem is much larger. When two WARCWriters are started at the same time, they COULD possibly have the exact same, writing to the same file. This is obviously bad. Solution incoming.

* Fix: check if file exists already.

Now, sadly, this isn't perfect, but I feel like this is a needed check.

* fix: use a mutex to ensure we are making a unique file per WARC writer.

* Fix: allow IPv6 addresses in WARC-IP-Address

After fixing my IPv6 on my main server, I noticed AT was defaulting to IPv6 for Telegram. Taking a look at the WARC, they were adding "WARC-IP-Address: 2001:67c:4e8:1033:4:100:0:a".

Per the WARC 1.1 specification: "<per section 2.2 of RFC4291>". Reading section 2.2.1 of RFC4291, we appear to be perfectly in spec with the IP addresses returned.

Double checking with the same URL with this change on Zeno, we get "WARC-IP-Address: 2001:67c:4e8:1033:4:100:0:a", the exact same, RFC4291 compliant IPv6 address.

In conclusion: we have the same behavior as wget-at and as far as I can tell, compliant with RFC4291 as requested by the WARC 1.1 spec.