DRAFT: Write WARC responses to temporary files to (hopefully) avoid OOM issues. (#6)
* Fix: WARC-Target-URI when used with HTTP/S proxies.

* fix: Panic at the end of Zeno run.

If a warc file doesn't have any records inside of it, it will panic, because there's no file on the filesystem. I currently don't believe this to be panic worthy, but it is obviously important that a file we expect to be there is missing.

* Fix: the old solution wasn't really a "fix" and the problem is much larger. When two WARCWriters are started at the same time, they COULD possibly have the exact same, writing to the same file. This is obviously bad. Solution incoming.

* Fix: check if file exists already.

Now, sadly, this isn't perfect, but I feel like this is a needed check.

* fix: use a mutex to ensure we are making a unique file per WARC writer.

* Fix: allow IPv6 addresses in WARC-IP-Address

After fixing my IPv6 on my main server, I noticed AT was defaulting to IPv6 for Telegram. Taking a look at the WARC, they were adding "WARC-IP-Address: 2001:67c:4e8:1033:4:100:0:a".

Per the WARC 1.1 specification: "<per section 2.2 of RFC4291>". Reading section 2.2.1 of RFC4291, we appear to be perfectly in spec with the IP addresses returned.

Double checking with the same URL with this change on Zeno, we get "WARC-IP-Address: 2001:67c:4e8:1033:4:100:0:a", the exact same, RFC4291 compliant IPv6 address.

In conclusion: we have the same behavior as wget-at and as far as I can tell, compliant with RFC4291 as requested by the WARC 1.1 spec.

* DRAFT: write temporary files to disk to hopefully avoid OOM

Add WARCTempDir to our custom HTTP client, currently automatically configured, but we could add some configuration options in the future.

Change GetSHA1FromFile to support the headers being added separately from the actual file.

dialer.go: copy a limited number of bytes, right now 10K, if we get exactly 10K, we probably hit a limit, from there we want to split out the header, create a temporary file, add the content we previously got and the rest of the pipe.

Change up how we are reading the HTTP response depending on how we are writing it to the WARC file (either file, or .Content).

Deduplication works in the exact same way, but we have to remove files when we no longer need them.

write.go:
Remove the temporary files after they're written with a defer. Check if we sent the headers through r.Content, if we did, read it all, then use it for various calculations like Content-Length and Block-Digest.

After that, we use those headers, if they exist, to write to the actual WARC file.

After testing and coding this for the last week, this _appears_ to work. The code is a bit rough, but I've played back sites with Transfer-Encoding, video files, images, etc. Everything I've seen so far has the correct Payload-Digest (block-digest is a bit harder to validate..., but seems correct), and has perfectly valid warcs (as far as I can tell!), while being written to temporary files.

* Fix: tests and ensuring directories exist to write the temporary files into.

Total oversight by me, Zeno was creating them before. This ensures everything exists.

* Fix: address Corentin's comments and switch to path.Join for temporary directory.

* Fix: change memory -> disk threshold to 2MB from 10KB