feat: add initial "read" testing on our dynamically written warcs
on every single HTTPClient test, look for all files for that test (since we can't really return the WARC file name) and check that WARC for the correct amount of response records and Payload-Digest. Once we've reached EOF, check that we read the expected number, if we haven't error out. If we have, all good, everything works!

This vastly improves our current tests, as we can ensure we're actually writing WARCs, and that they most likely hold the correct content.