Add: simple benchmark with two tests from above.
It's essentially two of the tests from above, but modified to allow for the go benchmarking.

From a quick test of "rm -rf warcs && go test -bench=. -count 5" it appears that the gzip change was quite large!

With a change of 8-9 to 38-40 in the Over2MB concurrent test and about a 50-100 improvement in Under2MB. Specific results published: https://gist.github.com/NGTmeaty/11258f9bce5ebf2ddfbeb8387636ea5a

(on a 4 core, 8GB ram machine.)