Fix: current tests
The issues appeared to be related to listening the older error channels. We can move all of these errors to the newer httpClient error channel.

We also add a new test to one of the major "SHA1 error" causes. Improper handling of WARC data. My going assumption is that HTTP errors that produce some response body are "half" copied into WARC, leaving an incomplete version, causing the SHA1 errors we see. This adds a test to confirm that behavior. The best result would be outputting specific errors, but I'm not sure we can specifically catch this, looking into that.

Fixing SHA1 will require work on Zeno (which is in-progress) (and other `warc` apps).