Improve HTTP Request line validation
Refactor `isLineStartingWithHTTPMethod` to include all the logic that is
necessary to validate an HTTP request (prefix and suffix).

Previously, some of the logic was in `isLineStartingWithHTTPMethod` and
some was in `dialer.go` where the function was used.

Rename it to `isHTTPRequest`. (IMHO `isHTTPRequestLine` is redundant as
`line` is the function param.

Add unit tests.