Improve the use of goleak in unit tests
Instead of using goleak in every unit test, run it in `TestMain` and it
applies to all unit tests.

The project didn't have a `TestMain` so I created a `main_test.go` for
that reason. `TestMain` could be added to any test file and it would
work. I thought that it would be clearer to make a separate
`main_test.go` file.

This method is also referenced in https://github.com/uber-go/goleak?tab=readme-ov-file#quick-start